Moving typedef struct to typedef spec. Adding usings for namespace spec and in piggy.  Untabifying. Converting array types to sequence of fields in struct. char in C converted to byte in C#.

+++ PiggyRuntime/Helper.cs
@@ -72,7 +72,7 @@ namespace PiggyRuntime
             {"float", "float"},
             {"double", "double"},
             {"bool", "bool"},
-            {"char", "int"},
+            {"char", "byte"},
         };
 
         public static string ModNonParamUsageType(string type)
